separate worksheet component
