--- frontend/src/views/Excel/excel.css
@@ -1,3 +1,0 @@
-button:focus {
-  outline: none
-}
